+++ src/tools/uid.ts
@@ -1,5 +1,5 @@
 import { nanoid } from 'nanoid';
 
 export function generateUID() {
-  return nanoid().replace(/(^\d+|-|_)/g, '');
+  return nanoid().replace(/^\d+/g, '').replace(/-/g, '_');
 }